API

SkewedAxisFourierAnalysis.build_sheared_function_from_sinousoids — Method

build_function_from_sinousoids(opdata, xx, yy, θ)

Builds the function from the sinousoids, on x and y axis skewed at an angle θ. The function is evaluated at the points xx and yy, which are the meshgrid points where you want the function to be evaluated.

Arguments

opdata: The sinousoids data. It is a vector of tuples, where each tuple is (i, j, mag, ang), where i and j are the harmonic indices, mag is the magnitude of the harmonic, and ang is the phase of the harmonic.

SkewedAxisFourierAnalysis.convert_image_to_data — Method

convert_image_to_data(img, cb, decimation_factor, lowerlimit, upperlimit)

SkewedAxisFourierAnalysis.create_image_from_data — Method

create_image_from_data(fval, custom_color_map)

Create an image from 2D data fval using the custom color map custom_color_map.

SkewedAxisFourierAnalysis.crop_transparent_border_of_image — Method

crop_transparent_border_of_image(warped_img)

Crop the transparent border of the image.

SkewedAxisFourierAnalysis.extract_dominant_harmonics — Method

extract_dominant_harmonics(image_data, threshold=0.005)

Extract the dominant harmonics from the image data (2D data). The threshold is used to determine the high power mask. The frequencies are extracted from the high power mask and centered around the zero frequency. frequencies and their Fourier coefficients are returned.

Arguments

image_data: The image data to be analyzed.

threshold=0.005: The threshold for high power.

SkewedAxisFourierAnalysis.fft_power_analysis — Method

fft_power_analysis(image_data, threshold=0.005)

Perform a power analysis on the Fourier transform of the image data. Based on the threshold, the high power mask is generated. Total power spectral density of frequencies in high power mask is greater than the threshold.

Arguments

image_data: The image data to be analyzed.

threshold=0.005: The threshold for high power.

Return value

high_power_mask: The mask for high power.

image_fft_raw: The raw Fourier transform of the image data.

SkewedAxisFourierAnalysis.filter_principal_frequencies_for_fft_of_real_data — Method

filter_principal_frequencies_for_fft_of_real_data(sinousoids)

Filter the principal frequencies for the FFT of real data. As FFT of real data have the property $F(\omega) = conj(F(-\omega))$ We can filter the principal frequencies for the FFT of real data and still fully reconstruct the image.

SkewedAxisFourierAnalysis.findnearest_value_index — Method

findnearest_value_index(f::F, x::AbstractArray{T}, x1::T) where {T,F<:Function}

Find the index of the nearest value in the array x to the value x1 using the function f to measure the distance between these two points.

SkewedAxisFourierAnalysis.map_image_to_data — Method

map_image_to_data(img::Matrix{RGB{T}}, cb_line::Vector{RGB{T}}, lowerlimit, upperlimit)

Map a scientific image to a data array using a colorbar line, and lower and upper limits of the colorbar.

Arguments

img::Matrix{RGB{T}}: The scientific image to be mapped to data. Element type is RGB{T}.

cb_line::Vector{RGB{T}}: The colorbar line. Element type is RGB{T}.

lowerlimit: The lower limit of the colorbar.

upperlimit: The upper limit of the colorbar.

julia imagedata = mapimagetodata(img, cb_line, lowerlimit, upperlimit)

SkewedAxisFourierAnalysis.rgb_to_rgba — Method

rgb_to_rgba(img::Array{RGB{T},2}) where {T<:Number}

Convert an RGB image to an RGBA image.

SkewedAxisFourierAnalysis.rgba_to_rgb — Method

rgba_to_rgb(img::Array{RGBA{T},2}) where {T<:Number}

Convert an RGBA image to an RGB image.

SkewedAxisFourierAnalysis.shear_image — Method

shear_image(img, θ)

Shear the image at an angle θ.

Arguments

img: The image to be sheared.

θ: The angle in degrees at which the image is to be sheared.

SkewedAxisFourierAnalysis.write_opdata_to_file — Method

write_opdata_to_file(opdata, sine_filename)

Write the opdata to a file with the format i, j, mag, ang. The last line is the DC component.